cleanup: remove unnecessary stores:
* remove the majority of files in 'web-client/src/lib/stores'
* move all the app-related stores to '/stores/app' and reexport from index
